Don't export autoprofile.NewTimer()
