+++ autoprofile/auto_profiler.go
@@ -3,7 +3,6 @@ package autoprofile
 import (
 	"path/filepath"
 	"sync/atomic"
-	"time"
 )
 
 const (
@@ -140,78 +139,6 @@ func (ap *autoProfiler) Disable() {
 	ap.blockSamplerScheduler.stop()
 
 	log.debug("profiler disabled")
-}
-
-func recoverAndLog() {
-	if err := recover(); err != nil {
-		log.error("recovered from panic in agent:", err)
-	}
-}
-
-type Timer struct {
-	delayTimer         *time.Timer
-	delayTimerDone     chan bool
-	intervalTicker     *time.Ticker
-	intervalTickerDone chan bool
-	stopped            bool
-}
-
-func NewTimer(delay time.Duration, interval time.Duration, job func()) *Timer {
-	t := &Timer{
-		stopped: false,
-	}
-
-	t.delayTimerDone = make(chan bool)
-	t.delayTimer = time.NewTimer(delay)
-	go func() {
-		defer recoverAndLog()
-
-		select {
-		case <-t.delayTimer.C:
-			if interval > 0 {
-				t.intervalTickerDone = make(chan bool)
-				t.intervalTicker = time.NewTicker(interval)
-				go func() {
-					defer recoverAndLog()
-
-					for {
-						select {
-						case <-t.intervalTicker.C:
-							job()
-						case <-t.intervalTickerDone:
-							return
-						}
-					}
-				}()
-			}
-
-			if delay > 0 {
-				job()
-			}
-		case <-t.delayTimerDone:
-			return
-		}
-	}()
-
-	return t
-}
-
-func (t *Timer) Stop() {
-	if !t.stopped {
-		t.stopped = true
-
-		t.delayTimer.Stop()
-		close(t.delayTimerDone)
-
-		if t.intervalTicker != nil {
-			t.intervalTicker.Stop()
-			close(t.intervalTickerDone)
-		}
-	}
-}
-
-func createTimer(delay time.Duration, interval time.Duration, job func()) *Timer {
-	return NewTimer(delay, interval, job)
 }
 
 type Flag struct {
